pyproject.toml
[project]
name = "tabular-eda"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "plotly>=6.3.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "ucimlrepo>=0.0.7",
    "umap-learn>=0.5.9.post2",
]
